[Refactor] Shifted some of model class into another folder.
