+++ src/main/java/controllers/temp/ConsoleInputController.java
@@ -15,13 +15,13 @@ import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Scanner;
-import models.commands.RescheduleCommand;
+import models.temp.commands.RescheduleCommand;
 import models.data.IProject;
-import models.tasks.IRecurring;
-import models.tasks.ITask;
-import models.tasks.PeriodTask;
-import models.tasks.Recurring;
-import models.tasks.TaskList;
+import models.temp.tasks.IRecurring;
+import models.temp.tasks.ITask;
+import models.temp.tasks.PeriodTask;
+import models.temp.tasks.Recurring;
+import models.temp.tasks.TaskList;
 import util.repositories.ProjectRepository;
 import views.CLIView;
