+++ src/main/java/models/temp/commands/ICommand.java
@@ -1,6 +1,6 @@
-package models.commands;
+package models.temp.commands;
 
-import models.tasks.TaskList;
+import models.temp.tasks.TaskList;
 
 public interface ICommand {
     void execute(TaskList taskList);
