+++ src/main/java/models/temp/commands/IDateSettable.java
@@ -1,4 +1,4 @@
-package models.commands;
+package models.temp.commands;
 
 public interface IDateSettable {
     void setDateTime(String newDateTime);
