+++ src/main/java/models/temp/commands/RescheduleCommand.java
@@ -1,7 +1,7 @@
-package models.commands;
+package models.temp.commands;
 
-import models.tasks.TaskList;
-import models.tasks.ToDos;
+import models.temp.tasks.TaskList;
+import models.temp.tasks.ToDos;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
