+++ src/main/java/models/temp/tasks/Recurring.java
@@ -1,4 +1,4 @@
-package models.tasks;
+package models.temp.tasks;
 
 import java.util.Date;
