+++ src/main/java/models/temp/tasks/TaskList.java
@@ -1,11 +1,10 @@
-package models.tasks;
+package models.temp.tasks;
 
 import java.io.Serializable;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.Date;
