+++ src/test/java/DukeTest.java
@@ -3,9 +3,9 @@ import exceptions.DukeException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import models.tasks.Deadline;
-import models.tasks.Event;
-import models.tasks.ToDos;
+import models.temp.tasks.Deadline;
+import models.temp.tasks.Event;
+import models.temp.tasks.ToDos;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
